Update shell scripts to install config files

+++ config/autostart/run.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
-ABSOLUTE_PARENT_PATH=$(realpath $(dirname $BASH_SOURCE))
-ln -svf $ABSOLUTE_PARENT_PATH -t ~/.config
+# Config polybar and redshift for startup if using xorg & xfce4
+if [ -n "$DISPLAY" ] || [ -S /tmp/.X11-unix/X0 ]; then
+    ABSOLUTE_PARENT_PATH=$(realpath $(dirname $BASH_SOURCE))
+    ln -svf $ABSOLUTE_PARENT_PATH -t ~/.config
+fi
