+++ config/rofi/run.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 
+if [ -n "$WAYLAND_DISPLAY" ] || [ -S /run/user/$(id -u)/wayland-0 ]; then
+    exit
+fi
+sudo pacman -S --needed --noconfirm rofi
 ABSOLUTE_PARENT_PATH=$(realpath $(dirname $BASH_SOURCE))
 ln -svf $ABSOLUTE_PARENT_PATH -t ~/.config
